SWATCH-5264 Refactor swatch-tally and swatch-contracts to use common test framework models
