+++ swatch-tally/ct/java/tests/BaseTallyComponentTest.java
@@ -32,6 +32,7 @@ import com.redhat.swatch.component.tests.api.HbiDatabase;
 import com.redhat.swatch.component.tests.api.KafkaBridge;
 import com.redhat.swatch.component.tests.api.KafkaBridgeService;
 import com.redhat.swatch.component.tests.api.SpringBoot;
+import com.redhat.swatch.component.tests.api.SubscriptionsAccessLevel;
 import com.redhat.swatch.component.tests.api.SwatchDatabase;
 import com.redhat.swatch.component.tests.api.Unleash;
 import com.redhat.swatch.component.tests.api.Wiremock;
@@ -44,7 +45,6 @@ import com.redhat.swatch.tally.test.model.InstanceData;
 import com.redhat.swatch.tally.test.model.InstanceResponse;
 import com.redhat.swatch.tally.test.model.TallyReportData;
 import com.redhat.swatch.tally.test.model.TallyReportDataPoint;
-import domain.SubscriptionsAccessLevel;
 import java.time.Duration;
 import java.time.OffsetDateTime;
 import java.util.HashMap;
